chore: apply the release audit
`make audit` before cutting a release. Kept to what it flags and the project
does not deliberately allow:

- sorted `__all__` in three modules
- `itertools.pairwise` where a `zip(x, x[1:])` was walking successive pairs
- dropped six `int()` casts on values that were already integers
- collapsed two nested `if`s
- en dashes and multiplication signs out of the docstrings, hyphen and x instead

Left alone, on purpose: S603 (every ffmpeg call is a subprocess), S311 (seeded
random for B-roll), ARG002 (the block interface), PERF203 (try/except in a loop
is what per-step error handling is), and the RUF100 `# noqa`s, which document
exactly what this audit reports.

`_check_block_params` remains at complexity 48, four times the threshold. It is
a per-block if-chain; splitting it is a real refactor, not an audit fix.

+++ src/lemontage/engine/timecode.py
@@ -46,7 +46,7 @@ def _parse_timecode(text: str) -> float:
 
 def to_timecode(seconds: float) -> str:
     """Format seconds as ``HH:MM:SS.mmm`` (for SRT and ffmpeg)."""
-    millis = int(round(seconds * 1000))
+    millis = round(seconds * 1000)
     hours, millis = divmod(millis, 3_600_000)
     minutes, millis = divmod(millis, 60_000)
     secs, millis = divmod(millis, 1000)
